Streamline LCHTC and EFDTC
- Revert method attribute disable removal in NaiveBayes-based learning nodes
