+++ src/skmultiflow/trees/nodes/lc_hoeffding_tree_classifier_nodes.py
@@ -1,0 +1,189 @@
+from skmultiflow.trees.attribute_observer import NumericAttributeClassObserverGaussian
+from skmultiflow.trees.attribute_observer import NominalAttributeClassObserver
+from skmultiflow.trees.nodes import ActiveLearningNode
+from skmultiflow.trees.nodes import InactiveLearningNode
+
+from skmultiflow.bayes import do_naive_bayes_prediction
+
+
+class LCActiveLearningNode(ActiveLearningNode):
+    """ Active Learning node for the Label Combination Hoeffding Tree.
+
+    Parameters
+    ----------
+    initial_class_observations: dict (class_value, weight) or None
+        Initial class observations
+
+    """
+    def __init__(self, initial_class_observations):
+        super().__init__(initial_class_observations)
+
+    def learn_from_instance(self, X, y, weight, ht):
+
+        if ht.leaf_prediction != ht._NAIVE_BAYES_ADAPTIVE:
+            y = ''.join(str(e) for e in y)
+            y = int(y, 2)
+
+        try:
+            self._observed_class_distribution[y] += weight
+        except KeyError:
+            self._observed_class_distribution[y] = weight
+
+        for i in range(len(X)):
+            try:
+                obs = self._attribute_observers[i]
+            except KeyError:
+                if ht.nominal_attributes is not None and i in ht.nominal_attributes:
+                    obs = NominalAttributeClassObserver()
+                else:
+                    obs = NumericAttributeClassObserverGaussian()
+                self._attribute_observers[i] = obs
+            obs.observe_attribute_class(X[i], int(y), weight)
+
+
+class LCInactiveLearningNode(InactiveLearningNode):
+    """ Inactive Learning node for the Label Combination Hoeffding Tree.
+
+    Parameters
+    ----------
+    initial_class_observations: dict (class_value, weight) or None
+        Initial class observations
+
+    """
+    def __init__(self, initial_class_observations=None):
+        """ LCInactiveLearningNode class constructor. """
+        super().__init__(initial_class_observations)
+
+    def learn_from_instance(self, X, y, weight, ht):
+
+        i = ''.join(str(e) for e in y)
+        i = int(i, 2)
+        try:
+            self._observed_class_distribution[i] += weight
+        except KeyError:
+            self._observed_class_distribution[i] = weight
+
+
+class LCLearningNodeNB(LCActiveLearningNode):
+    """ Learning node for the Label Combination Hoeffding Tree that uses Naive
+    Bayes models.
+
+    Parameters
+    ----------
+    initial_class_observations: dict (class_value, weight) or None
+        Initial class observations
+
+    """
+    def __init__(self, initial_class_observations):
+        """ LCLearningNodeNB class constructor. """
+        super().__init__(initial_class_observations)
+
+    def get_class_votes(self, X, ht):
+        """Get the votes per class for a given instance.
+
+        Parameters
+        ----------
+        X: numpy.ndarray of length equal to the number of features.
+            Instance attributes.
+        ht: HoeffdingTreeClassifier
+            Hoeffding Tree.
+
+        Returns
+        -------
+        dict (class_value, weight)
+            Class votes for the given instance.
+
+        """
+        if self.get_weight_seen() >= ht.nb_threshold:
+            return do_naive_bayes_prediction(
+                X, self._observed_class_distribution, self._attribute_observers
+            )
+        else:
+            return super().get_class_votes(X, ht)
+
+    def disable_attribute(self, att_index):
+        """ Disable an attribute observer.
+
+        Disabled in Nodes using Naive Bayes, since poor attributes are used in
+        Naive Bayes calculation.
+
+        Parameters
+        ----------
+        att_index: int
+            Attribute index.
+
+        """
+        pass
+
+
+class LCLearningNodeNBA(LCLearningNodeNB):
+    """ Learning node for the Label Combination Hoeffding Tree that uses
+    Adaptive Naive Bayes models.
+
+    Parameters
+    ----------
+    initial_class_observations: dict (class_value, weight) or None
+        Initial class observations
+
+    """
+    def __init__(self, initial_class_observations):
+        """LCLearningNodeNBA class constructor. """
+        super().__init__(initial_class_observations)
+        self._mc_correct_weight = 0.0
+        self._nb_correct_weight = 0.0
+
+    def learn_from_instance(self, X, y, weight, ht):
+        """Update the node with the provided instance.
+
+        Parameters
+        ----------
+        X: numpy.ndarray of length equal to the number of features.
+            Instance attributes for updating the node.
+        y: int
+            Instance class.
+        weight: float
+            The instance's weight.
+        ht: HoeffdingTreeClassifier
+            The Hoeffding Tree to update.
+
+        """
+
+        y = ''.join(str(e) for e in y)
+        y = int(y, 2)
+
+        if self._observed_class_distribution == {}:
+            # All target_values equal, default to class 0
+            if 0 == y:
+                self._mc_correct_weight += weight
+        elif max(self._observed_class_distribution,
+                 key=self._observed_class_distribution.get) == y:
+            self._mc_correct_weight += weight
+        nb_prediction = do_naive_bayes_prediction(
+            X, self._observed_class_distribution, self._attribute_observers
+        )
+        if max(nb_prediction, key=nb_prediction.get) == y:
+            self._nb_correct_weight += weight
+
+        super().learn_from_instance(X, y, weight, ht)
+
+    def get_class_votes(self, X, ht):
+        """Get the votes per class for a given instance.
+
+        Parameters
+        ----------
+        X: numpy.ndarray of length equal to the number of features.
+            Instance attributes.
+        ht: HoeffdingTreeClassifier
+            Hoeffding Tree.
+
+        Returns
+        -------
+        dict (class_value, weight)
+            Class votes for the given instance.
+
+        """
+        if self._mc_correct_weight > self._nb_correct_weight:
+            return self._observed_class_distribution
+        return do_naive_bayes_prediction(
+            X, self._observed_class_distribution, self._attribute_observers
+        )
